Update README.md
This is a clean, responsive personal portfolio website built using **HTML** and **CSS** to showcase my skills, projects, and profile as a developer. It includes sections like **About Me**, **Skills**, **Projects**, and **Contact**, all crafted with a focus on simplicity, design, and usability. The site is fully mobile-friendly and serves as an interactive resume to present my work professionally.